Decouple  (#22)
* Improvements, maintenance and decoupling

* Keep nested plugin tests on the current Hyprland ABI

The nested helper was still depending on a repo-root hyprexpo.so and direct plugin dispatchers in generated config. That made local testing drift from the requested TARGET build and caused Hyprland to reject plugin dispatchers before the plugin had registered them. Hyprland 0.55.1 also requires the pluginAPIVersion export before it will call pluginInit.

The helpers now build/load a user-owned cached artifact, route test binds through hyprctl, and include basic nested-session controls for spawning terminals and moving across workspaces.

Constraint: Hyprland 0.55.1 requires pluginAPIVersion before pluginInit
Rejected: Use repo-root hyprexpo.so | it can be stale or root-owned and masks the actual build under test
Confidence: high
Scope-risk: narrow
Directive: Keep nested helper builds user-owned and loaded from the same path the generated config references
Tested: bash -n scripts/run-nested.sh scripts/dev-watch.sh scripts/dev-link.sh
Tested: make test
Tested: make all TARGET=/tmp/hyprexpo.so
Tested: timeout --kill-after=2s 10s scripts/run-nested.sh; latest nested log loaded plugin hyprexpo and unloaded cleanly
Not-tested: Manual F10 and keyboard selection interaction inside a long-lived nested window

---------
