+++ .github/workflows/release.yml
@@ -34,17 +34,42 @@ jobs:
             pixman \
             wayland \
             libxkbcommon \
+            lua \
             pkgconf
 
       - name: Checkout
         uses: actions/checkout@v4
 
+      - name: Verify parser tests
+        run: make test
+
       - name: Build plugin
         run: make all
+
+      - name: Record build provenance
+        run: |
+          {
+            echo "hyprctl version:"
+            hyprctl version || true
+            echo
+            echo "pkg-config hyprland:"
+            pkg-config --modversion hyprland || true
+            echo
+            echo "pkg-config lua5.4:"
+            pkg-config --modversion lua5.4 || true
+            echo
+            echo "compiler:"
+            c++ --version | head -n 1
+            echo
+            echo "ldd -r hyprexpo.so:"
+            ldd -r hyprexpo.so || true
+          } | tee release-provenance.txt
 
       - name: Create release
         if: startsWith(github.ref, 'refs/tags/')
         uses: softprops/action-gh-release@v2
         with:
-          files: hyprexpo.so
+          files: |
+            hyprexpo.so
+            release-provenance.txt
           generate_release_notes: true
