+++ Dispatchers.hpp
@@ -1,0 +1,14 @@
+#pragma once
+
+#define WLR_USE_UNSTABLE
+
+#include "globals.hpp"
+#include <hyprland/src/config/values/types/StringValue.hpp>
+#include <hyprland/src/devices/IKeyboard.hpp>
+
+SP<Config::Values::CStringValue> createCancelKeyConfig();
+void                             resetDispatcherRuntime();
+void                             registerHyprexpoDispatchers();
+bool                             isRenderingOverview();
+bool                             shouldCancelOverview(const IKeyboard::SKeyEvent& event);
+bool                             shouldSelectWorkspaceFromKey(const IKeyboard::SKeyEvent& event);
