+++ OverviewPassElement.cpp
@@ -1,12 +1,15 @@
 #include "OverviewPassElement.hpp"
 #include <hyprland/src/render/OpenGL.hpp>
-#include "overview.hpp"
+#include "Overview.hpp"
 
 COverviewPassElement::COverviewPassElement() {
     ;
 }
 
 std::vector<UP<IPassElement>> COverviewPassElement::draw() {
+    if (!g_pOverview)
+        return {};
+
     g_pOverview->fullRender();
     return {};
 }
@@ -20,14 +23,14 @@ bool COverviewPassElement::needsPrecomputeBlur() {
 }
 
 std::optional<CBox> COverviewPassElement::boundingBox() {
-    if (!g_pOverview->pMonitor)
+    if (!g_pOverview || !g_pOverview->pMonitor)
         return std::nullopt;
 
     return CBox{{}, g_pOverview->pMonitor->m_size};
 }
 
 CRegion COverviewPassElement::opaqueRegion() {
-    if (!g_pOverview->pMonitor)
+    if (!g_pOverview || !g_pOverview->pMonitor)
         return CRegion{};
 
     return CBox{{}, g_pOverview->pMonitor->m_size};
