--- overview.cpp
@@ -1,1631 +1,0 @@
-#include "overview.hpp"
-#include <any>
-#include <map>
-#include <hyprland/src/event/EventBus.hpp>
-#define private   public
-#define protected public
-#include <hyprland/src/render/Renderer.hpp>
-#include <hyprland/src/Compositor.hpp>
-#include <hyprland/src/config/ConfigValue.hpp>
-#include <hyprland/src/config/ConfigManager.hpp>
-#include <hyprland/src/config/shared/actions/ConfigActions.hpp>
-#include <hyprland/src/config/shared/animation/AnimationTree.hpp>
-#include <hyprland/src/managers/animation/AnimationManager.hpp>
-#include <hyprland/src/managers/animation/DesktopAnimationManager.hpp>
-#include <hyprland/src/managers/input/InputManager.hpp>
-#include <hyprland/src/managers/PointerManager.hpp>
-#include <hyprland/src/helpers/time/Time.hpp>
-#include <hyprland/src/helpers/varlist/VarList.hpp>
-#include <hyprland/src/helpers/Format.hpp>
-#include <drm_fourcc.h>
-#undef private
-#undef protected
-#include "OverviewPassElement.hpp"
-#include <hyprland/src/render/OpenGL.hpp>
-#include <hyprland/src/config/shared/complex/ComplexDataTypes.hpp>
-#include <pango/pangocairo.h>
-#include <algorithm>
-#include <cctype>
-#include <cmath>
-#include <string>
-#include <vector>
-
-namespace CompatHyprlandAPI {
-struct SConfigValueCompat {
-    Config::INTEGER integer = 0;
-    Config::INTEGER* integerPtr = &integer;
-    Config::FLOAT floating = 0;
-    Config::FLOAT* floatingPtr = &floating;
-    Config::STRING string;
-    void* ptr = nullptr;
-
-    void* getDataStaticPtr() {
-        return ptr;
-    }
-};
-
-static bool isStringConfig(const std::string& name) {
-    static const std::map<std::string, bool> STRING_CONFIGS = {
-        {"plugin:hyprexpo:workspace_method", true},
-        {"plugin:hyprexpo:border_color", true},
-        {"plugin:hyprexpo:border_color_current", true},
-        {"plugin:hyprexpo:border_color_focus", true},
-        {"plugin:hyprexpo:border_color_hover", true},
-        {"plugin:hyprexpo:border_style", true},
-        {"plugin:hyprexpo:label_text_mode", true},
-        {"plugin:hyprexpo:label_token_map", true},
-        {"plugin:hyprexpo:label_position", true},
-        {"plugin:hyprexpo:selection_label_token_map", true},
-        {"plugin:hyprexpo:selection_label_position", true},
-        {"plugin:hyprexpo:label_show", true},
-        {"plugin:hyprexpo:label_bg_shape", true},
-        {"plugin:hyprexpo:label_font_family", true},
-        {"plugin:hyprexpo:cancel_key", true},
-        {"plugin:hyprexpo:border_grad_current", true},
-        {"plugin:hyprexpo:border_grad_focus", true},
-        {"plugin:hyprexpo:border_grad_hover", true},
-    };
-
-    return STRING_CONFIGS.contains(name);
-}
-
-static bool isFloatConfig(const std::string& name) {
-    return name == "plugin:hyprexpo:tile_rounding_power" ||
-           name == "plugin:hyprexpo:label_scale_hover" ||
-           name == "plugin:hyprexpo:label_scale_focus";
-}
-
-static Config::STRING stringDefault(const std::string& name) {
-    static const std::map<std::string, Config::STRING> DEFAULTS = {
-        {"plugin:hyprexpo:workspace_method", "center current"},
-        {"plugin:hyprexpo:border_color", ""},
-        {"plugin:hyprexpo:border_color_current", "rgb(66ccff)"},
-        {"plugin:hyprexpo:border_color_focus", "rgb(ffcc66)"},
-        {"plugin:hyprexpo:border_color_hover", "rgb(aabbcc)"},
-        {"plugin:hyprexpo:border_style", "simple"},
-        {"plugin:hyprexpo:label_text_mode", "token"},
-        {"plugin:hyprexpo:label_token_map", ""},
-        {"plugin:hyprexpo:label_position", "center"},
-        {"plugin:hyprexpo:selection_label_token_map", "a,s,d,f,g,q,w,e,r,t,z,x,c,v,b"},
-        {"plugin:hyprexpo:selection_label_position", "top-right"},
-        {"plugin:hyprexpo:label_show", "always"},
-        {"plugin:hyprexpo:label_bg_shape", "circle"},
-        {"plugin:hyprexpo:label_font_family", "sans"},
-        {"plugin:hyprexpo:cancel_key", "escape"},
-        {"plugin:hyprexpo:border_grad_current", ""},
-        {"plugin:hyprexpo:border_grad_focus", ""},
-        {"plugin:hyprexpo:border_grad_hover", ""},
-    };
-
-    if (const auto it = DEFAULTS.find(name); it != DEFAULTS.end())
-        return it->second;
-    return "";
-}
-
-static Config::FLOAT floatDefault(const std::string& name) {
-    if (name == "plugin:hyprexpo:tile_rounding_power")
-        return 2.0F;
-    if (name == "plugin:hyprexpo:label_scale_hover" ||
-        name == "plugin:hyprexpo:label_scale_focus")
-        return 1.0F;
-    return 0.0F;
-}
-
-static Config::INTEGER intDefault(const std::string& name) {
-    static const std::map<std::string, Config::INTEGER> DEFAULTS = {
-        {"plugin:hyprexpo:columns", 3}, {"plugin:hyprexpo:gaps_in", 5},
-        {"plugin:hyprexpo:bg_col", 0xFF111111}, {"plugin:hyprexpo:gesture_distance", 200}, {"plugin:hyprexpo:show_cursor", 1},
-        {"plugin:hyprexpo:keynav_enable", 1}, {"plugin:hyprexpo:border_width", 2},
-        {"plugin:hyprexpo:label_enable", 1}, {"plugin:hyprexpo:label_color", 0xFFFFFFFF},
-        {"plugin:hyprexpo:label_font_size", 16}, {"plugin:hyprexpo:label_color_default", 0xFFFFFFFF},
-        {"plugin:hyprexpo:label_color_hover", 0xFFEEEEEE}, {"plugin:hyprexpo:label_color_focus", 0xFFFFCC66},
-        {"plugin:hyprexpo:label_color_current", 0xFF66CCFF}, {"plugin:hyprexpo:label_bg_enable", 1},
-        {"plugin:hyprexpo:label_bg_color", 0x88000000}, {"plugin:hyprexpo:label_bg_rounding", 8},
-        {"plugin:hyprexpo:label_padding", 8}, {"plugin:hyprexpo:label_pixel_snap", 1},
-        {"plugin:hyprexpo:selection_label_enable", 0}, {"plugin:hyprexpo:selection_label_offset_x", 6},
-        {"plugin:hyprexpo:selection_label_offset_y", 6}, {"plugin:hyprexpo:selection_label_color", 0xFFFFCC66},
-        {"plugin:hyprexpo:keynav_wrap_h", 1}, {"plugin:hyprexpo:keynav_wrap_v", 1},
-    };
-
-    if (const auto it = DEFAULTS.find(name); it != DEFAULTS.end())
-        return it->second;
-    return 0;
-}
-
-static SConfigValueCompat* getConfigValue(HANDLE, const std::string& name) {
-    static std::map<std::string, SConfigValueCompat> VALUES;
-    auto& compat = VALUES[name];
-
-    if (isStringConfig(name)) {
-        compat.string = stringDefault(name);
-        compat.ptr = &compat.string;
-    } else if (isFloatConfig(name)) {
-        compat.floating = floatDefault(name);
-        compat.floatingPtr = &compat.floating;
-        compat.ptr = &compat.floatingPtr;
-    } else {
-        compat.integer = intDefault(name);
-        compat.integerPtr = &compat.integer;
-        compat.ptr = &compat.integerPtr;
-    }
-
-    const auto VALUE = Config::mgr()->getConfigValue(name);
-    if (VALUE.dataptr && VALUE.type && *VALUE.type == typeid(Config::STRING)) {
-        auto* ptr = reinterpret_cast<Config::STRING* const*>(VALUE.dataptr);
-        compat.ptr = ptr && *ptr ? *ptr : &compat.string;
-        return &compat;
-    }
-
-    if (VALUE.dataptr) {
-        compat.ptr = const_cast<void*>(static_cast<const void*>(VALUE.dataptr));
-        return &compat;
-    }
-
-    return &compat;
-}
-}
-
-#define HyprlandAPI CompatHyprlandAPI
-
-static void clearWithColor(const CHyprColor& color) {
-    glClearColor(color.r, color.g, color.b, color.a);
-    glClear(GL_COLOR_BUFFER_BIT);
-}
-
-static uint32_t framebufferFormatWithAlpha(uint32_t drmFormat) {
-    const auto alphaFormat = NFormatUtils::alphaFormat(drmFormat);
-    return alphaFormat == 0 ? DRM_FORMAT_ABGR8888 : alphaFormat;
-}
-
-static bool isTransformRotated(wl_output_transform t) {
-    return t == WL_OUTPUT_TRANSFORM_90 || t == WL_OUTPUT_TRANSFORM_270 ||
-           t == WL_OUTPUT_TRANSFORM_FLIPPED_90 || t == WL_OUTPUT_TRANSFORM_FLIPPED_270;
-}
-
-static std::string trimString(std::string value) {
-    while (!value.empty() && std::isspace((unsigned char)value.front()))
-        value.erase(value.begin());
-    while (!value.empty() && std::isspace((unsigned char)value.back()))
-        value.pop_back();
-    return value;
-}
-
-static std::vector<std::string> splitCommaList(const std::string& value) {
-    std::vector<std::string> entries;
-    size_t                   start = 0;
-
-    while (start <= value.size()) {
-        size_t comma = value.find(',', start);
-        if (comma == std::string::npos)
-            comma = value.size();
-
-        entries.push_back(trimString(value.substr(start, comma - start)));
-
-        if (comma == value.size())
-            break;
-        start = comma + 1;
-    }
-
-    return entries;
-}
-
-static std::string lowerString(std::string value) {
-    std::transform(value.begin(), value.end(), value.begin(), [](unsigned char c) { return std::tolower(c); });
-    return value;
-}
-
-static std::string fallbackTokenForVisibleIndex(int visibleIndex) {
-    if (visibleIndex < 9)
-        return std::to_string(visibleIndex + 1);
-    if (visibleIndex == 9)
-        return "0";
-    if (visibleIndex < 36)
-        return std::string(1, char('a' + visibleIndex - 10));
-
-    return "";
-}
-
-static int fallbackTokenToVisibleIndex(const std::string& token) {
-    if (token.size() != 1)
-        return -1;
-
-    const char c = token[0];
-    if (c >= '1' && c <= '9')
-        return c - '1';
-    if (c == '0')
-        return 9;
-    if (c >= 'a' && c <= 'z')
-        return 10 + c - 'a';
-    if (c >= 'A' && c <= 'Z')
-        return 10 + c - 'A';
-
-    return -1;
-}
-
-struct SHyprGradientSpec {
-    CHyprColor c1;
-    CHyprColor c2;
-    float      angleDeg = 0.f;
-    bool       valid    = false;
-};
-
-static bool parseHexRGBA8(const std::string& s, CHyprColor& out) {
-    // expects 8 hex digits RRGGBBAA
-    if (s.size() != 8)
-        return false;
-    auto hexTo = [](char c) -> int {
-        if (c >= '0' && c <= '9') return c - '0';
-        if (c >= 'a' && c <= 'f') return 10 + (c - 'a');
-        if (c >= 'A' && c <= 'F') return 10 + (c - 'A');
-        return -1;
-    };
-    auto byteAt = [&](int i) -> int {
-        int a = hexTo(s[i]);
-        int b = hexTo(s[i+1]);
-        if (a < 0 || b < 0) return -1;
-        return (a << 4) | b;
-    };
-    int r = byteAt(0);
-    int g = byteAt(2);
-    int b = byteAt(4);
-    int a = byteAt(6);
-    if (r < 0 || g < 0 || b < 0 || a < 0) return false;
-    out = CHyprColor{r / 255.f, g / 255.f, b / 255.f, a / 255.f};
-    return true;
-}
-
-static SHyprGradientSpec parseGradientSpec(const std::string& inRaw) {
-    // Accept forms like: "rgba(33ccffee) rgba(00ff99ee) 45deg"
-    // Extract 8 hex digits from two rgba(...) groups and an integer angle
-    SHyprGradientSpec spec;
-    std::string       s = inRaw;
-    // remove commas
-    s.erase(std::remove(s.begin(), s.end(), ','), s.end());
-    // find 1st rgba(XXXXXXXX)
-    auto p1 = s.find("rgba(");
-    auto p2 = s.find("rgba(", p1 == std::string::npos ? 0 : p1 + 1);
-    if (p1 == std::string::npos || p2 == std::string::npos)
-        return spec;
-    auto e1 = s.find(')', p1);
-    auto e2 = s.find(')', p2);
-    if (e1 == std::string::npos || e2 == std::string::npos)
-        return spec;
-    const std::string hex1 = s.substr(p1 + 5, e1 - (p1 + 5));
-    const std::string hex2 = s.substr(p2 + 5, e2 - (p2 + 5));
-    CHyprColor        c1, c2;
-    if (!parseHexRGBA8(hex1, c1) || !parseHexRGBA8(hex2, c2))
-        return spec;
-    // find angle
-    float angle = 0.f;
-    auto  pd    = s.find("deg", e2);
-    if (pd != std::string::npos) {
-        // collect digits before 'deg'
-        size_t beg = s.rfind(' ', pd);
-        if (beg == std::string::npos)
-            beg = e2 + 1;
-        try {
-            angle = std::stof(s.substr(beg, pd - beg));
-        } catch (...) { angle = 0.f; }
-    }
-    spec.c1       = c1;
-    spec.c2       = c2;
-    spec.angleDeg = angle;
-    spec.valid    = true;
-    return spec;
-}
-
-// Helper to detect if a border config string is a gradient or solid color
-static bool isGradientBorderSpec(const std::string& borderSpec) {
-    if (borderSpec.empty())
-        return false;
-    // Check if it contains gradient pattern: rgba(...) rgba(...) deg
-    return borderSpec.find("rgba(") != std::string::npos &&
-           borderSpec.rfind("rgba(") != borderSpec.find("rgba(");
-}
-
-static SP<Render::ITexture> renderNumberTexture(const std::string& text, const CHyprColor& color, const Vector2D& bufferSize, const float scale, const int fontSize) {
-    const auto CAIROSURFACE = cairo_image_surface_create(CAIRO_FORMAT_ARGB32, bufferSize.x, bufferSize.y);
-    const auto CAIRO        = cairo_create(CAIROSURFACE);
-
-    cairo_save(CAIRO);
-    cairo_set_operator(CAIRO, CAIRO_OPERATOR_CLEAR);
-    cairo_paint(CAIRO);
-    cairo_restore(CAIRO);
-
-    PangoLayout* layout = pango_cairo_create_layout(CAIRO);
-    pango_layout_set_text(layout, text.c_str(), -1);
-
-    // font options from config
-    static auto* const PFONTFAM = (Hyprlang::STRING const*)HyprlandAPI::getConfigValue(PHANDLE, "plugin:hyprexpo:label_font_family")->getDataStaticPtr();
-    static auto* const PFONTB   = (Hyprlang::INT* const*)HyprlandAPI::getConfigValue(PHANDLE, "plugin:hyprexpo:label_font_bold")->getDataStaticPtr();
-    static auto* const PFONTI   = (Hyprlang::INT* const*)HyprlandAPI::getConfigValue(PHANDLE, "plugin:hyprexpo:label_font_italic")->getDataStaticPtr();
-    static auto* const PTUNDER  = (Hyprlang::INT* const*)HyprlandAPI::getConfigValue(PHANDLE, "plugin:hyprexpo:label_text_underline")->getDataStaticPtr();
-    static auto* const PTSTRIKE = (Hyprlang::INT* const*)HyprlandAPI::getConfigValue(PHANDLE, "plugin:hyprexpo:label_text_strikethrough")->getDataStaticPtr();
-
-    PangoFontDescription* fontDesc = pango_font_description_from_string(*PFONTFAM);
-    pango_font_description_set_size(fontDesc, fontSize * scale * PANGO_SCALE);
-    pango_font_description_set_weight(fontDesc, **PFONTB ? PANGO_WEIGHT_BOLD : PANGO_WEIGHT_NORMAL);
-    pango_font_description_set_style(fontDesc, **PFONTI ? PANGO_STYLE_ITALIC : PANGO_STYLE_NORMAL);
-    pango_layout_set_font_description(layout, fontDesc);
-    pango_font_description_free(fontDesc);
-
-    if (**PTUNDER || **PTSTRIKE) {
-        PangoAttrList* attrs = pango_attr_list_new();
-        if (**PTUNDER) {
-            pango_attr_list_insert(attrs, pango_attr_underline_new(PANGO_UNDERLINE_SINGLE));
-        }
-        if (**PTSTRIKE) {
-            pango_attr_list_insert(attrs, pango_attr_strikethrough_new(TRUE));
-        }
-        pango_layout_set_attributes(layout, attrs);
-        pango_attr_list_unref(attrs);
-    }
-
-    pango_layout_set_width(layout, bufferSize.x * PANGO_SCALE);
-    pango_layout_set_ellipsize(layout, PANGO_ELLIPSIZE_NONE);
-
-    cairo_set_source_rgba(CAIRO, color.r, color.g, color.b, color.a);
-
-    PangoRectangle ink_rect, logical_rect;
-    pango_layout_get_extents(layout, &ink_rect, &logical_rect);
-
-    // center inside the provided buffer using ink rect (accounts for glyph bearings)
-    const int    inkW   = std::max(0, ink_rect.width / PANGO_SCALE);
-    const int    inkH   = std::max(0, ink_rect.height / PANGO_SCALE);
-    const int    inkX   = ink_rect.x / PANGO_SCALE; // can be negative
-    const int    inkY   = ink_rect.y / PANGO_SCALE; // can be negative
-    static auto* const* PCENTERADJX = (Hyprlang::INT* const*)HyprlandAPI::getConfigValue(PHANDLE, "plugin:hyprexpo:label_center_adjust_x")->getDataStaticPtr();
-    static auto* const* PCENTERADJY = (Hyprlang::INT* const*)HyprlandAPI::getConfigValue(PHANDLE, "plugin:hyprexpo:label_center_adjust_y")->getDataStaticPtr();
-    const double xOffset = (bufferSize.x - inkW) / 2.0 - inkX + **PCENTERADJX;
-    const double yOffset = (bufferSize.y - inkH) / 2.0 - inkY + **PCENTERADJY;
-
-    cairo_move_to(CAIRO, xOffset, yOffset);
-    pango_cairo_show_layout(CAIRO, layout);
-    g_object_unref(layout);
-
-    cairo_surface_flush(CAIROSURFACE);
-
-    auto tex = g_pHyprRenderer->createTexture(CAIROSURFACE);
-    if (tex) {
-        tex->setTexParameter(GL_TEXTURE_MAG_FILTER, GL_NEAREST);
-        tex->setTexParameter(GL_TEXTURE_MIN_FILTER, GL_NEAREST);
-    }
-
-    cairo_destroy(CAIRO);
-    cairo_surface_destroy(CAIROSURFACE);
-
-    return tex;
-}
-
-static void damageMonitor(WP<Hyprutils::Animation::CBaseAnimatedVariable> thisptr) {
-    g_pOverview->damage();
-}
-
-struct SWorkspacePreviewState {
-    bool     visible        = false;
-    bool     forceRendering = false;
-    float    alphaValue     = 1.F;
-    float    alphaGoal      = 1.F;
-    Vector2D offsetValue;
-    Vector2D offsetGoal;
-};
-
-static SWorkspacePreviewState applyWorkspacePreviewState(const PHLWORKSPACE& workspace) {
-    SWorkspacePreviewState state;
-    if (!workspace)
-        return state;
-
-    state.visible        = workspace->m_visible;
-    state.forceRendering = workspace->m_forceRendering;
-    state.alphaValue     = workspace->m_alpha->value();
-    state.alphaGoal      = workspace->m_alpha->goal();
-    state.offsetValue    = workspace->m_renderOffset->value();
-    state.offsetGoal     = workspace->m_renderOffset->goal();
-
-    workspace->m_visible        = true;
-    workspace->m_forceRendering = true;
-    workspace->m_alpha->setValueAndWarp(1.F);
-    *workspace->m_alpha = 1.F;
-    workspace->m_renderOffset->setValueAndWarp(Vector2D{});
-    *workspace->m_renderOffset = Vector2D{};
-
-    return state;
-}
-
-static void restoreWorkspacePreviewState(const PHLWORKSPACE& workspace, const SWorkspacePreviewState& state) {
-    if (!workspace)
-        return;
-
-    workspace->m_visible        = state.visible;
-    workspace->m_forceRendering = state.forceRendering;
-    workspace->m_alpha->setValueAndWarp(state.alphaValue);
-    *workspace->m_alpha = state.alphaGoal;
-    workspace->m_renderOffset->setValueAndWarp(state.offsetValue);
-    *workspace->m_renderOffset = state.offsetGoal;
-}
-
-static void removeOverview(WP<Hyprutils::Animation::CBaseAnimatedVariable> thisptr) {
-    g_pOverview.reset();
-}
-
-static bool shouldShowCursorDuringOverview() {
-    static auto* const* PSHOWCURSOR = (Hyprlang::INT* const*)HyprlandAPI::getConfigValue(PHANDLE, "plugin:hyprexpo:show_cursor")->getDataStaticPtr();
-    return **PSHOWCURSOR;
-}
-
-static void ensureOverviewCursorVisible(bool refreshPosition = false) {
-    if (!shouldShowCursorDuringOverview())
-        return;
-
-    g_pHyprRenderer->setCursorHidden(false);
-    g_pPointerManager->resetCursorImage();
-    if (refreshPosition)
-        g_pInputManager->simulateMouseMovement();
-}
-
-// Get workspace method configuration for a specific monitor
-// Returns pair of {isCenter, startWorkspaceID}
-static std::pair<bool, int> getWorkspaceMethodForMonitor(PHLMONITOR monitor) {
-    static auto const* PMETHOD = (Hyprlang::STRING const*)HyprlandAPI::getConfigValue(PHANDLE, "plugin:hyprexpo:workspace_method")->getDataStaticPtr();
-
-    const std::string monitorName = monitor->m_name;
-    const std::string configStr = std::string{*PMETHOD};
-
-    std::string methodStr;
-    bool foundMonitorConfig = false;
-
-    // Parse plugin config value with delimiter support:
-    // 1. Global: "center current" or "first 1"
-    // 2. Per-monitor: "DP-1 first 1, HDMI-1 center current"
-    // 3. Mixed: "DP-1 first 1, center current"
-    std::vector<std::string> entries;
-    size_t start = 0;
-    while (start < configStr.size()) {
-        size_t commaPos = configStr.find(',', start);
-        if (commaPos == std::string::npos)
-            commaPos = configStr.size();
-
-        std::string entry = configStr.substr(start, commaPos - start);
-        size_t firstNonSpace = entry.find_first_not_of(" \t");
-        size_t lastNonSpace = entry.find_last_not_of(" \t");
-        if (firstNonSpace != std::string::npos)
-            entry = entry.substr(firstNonSpace, lastNonSpace - firstNonSpace + 1);
-
-        if (!entry.empty())
-            entries.push_back(entry);
-
-        start = commaPos + 1;
-    }
-
-    std::string globalFallback;
-    for (const auto& entry : entries) {
-        CVarList tokens{entry, 0, 's', true};
-
-        if (tokens.size() == 3) {
-            const std::string entryMonitor = std::string{tokens[0]};
-            if (entryMonitor == monitorName) {
-                methodStr = std::string{tokens[1]} + " " + std::string{tokens[2]};
-                foundMonitorConfig = true;
-                break;
-            }
-        } else if (tokens.size() == 2 && globalFallback.empty()) {
-            globalFallback = entry;
-        }
-    }
-
-    if (!foundMonitorConfig) {
-        if (!globalFallback.empty())
-            methodStr = globalFallback;
-        else
-            methodStr = configStr;
-    }
-
-    // Parse the method string (format: "method workspace")
-    bool methodCenter = true;
-    int methodStartID = monitor->activeWorkspaceID();
-
-    CVarList method{methodStr, 0, 's', true};
-    if (method.size() >= 2) {
-        methodCenter = method[0] == "center";
-        methodStartID = getWorkspaceIDNameFromString(method[1]).id;
-        if (methodStartID == WORKSPACE_INVALID)
-            methodStartID = monitor->activeWorkspaceID();
-    } else if (method.size() > 0) {
-        Log::logger->log(Log::ERR, "[hyprexpo] invalid workspace_method for monitor {}: {}", monitorName, methodStr);
-    }
-
-    return {methodCenter, methodStartID};
-}
-
-COverview::~COverview() {
-    Render::GL::g_pHyprOpenGL->makeEGLCurrent();
-    images.clear(); // otherwise we get a vram leak
-    ensureOverviewCursorVisible(true);
-    if (pMonitor)
-        pMonitor->m_blurFBDirty = true;
-    resetSubmapIfNeeded();
-}
-
-COverview::COverview(PHLWORKSPACE startedOn_, bool swipe_) : startedOn(startedOn_), swipe(swipe_) {
-    const auto PMONITOR = g_pCompositor->getMonitorFromCursor();
-    pMonitor            = PMONITOR;
-
-    static auto* const* PCOLUMNS = (Hyprlang::INT* const*)HyprlandAPI::getConfigValue(PHANDLE, "plugin:hyprexpo:columns")->getDataStaticPtr();
-    static auto* const* PGAPS    = (Hyprlang::INT* const*)HyprlandAPI::getConfigValue(PHANDLE, "plugin:hyprexpo:gaps_in")->getDataStaticPtr();
-    static auto* const* PCOL     = (Hyprlang::INT* const*)HyprlandAPI::getConfigValue(PHANDLE, "plugin:hyprexpo:bg_col")->getDataStaticPtr();
-    static auto* const* PSKIP    = (Hyprlang::INT* const*)HyprlandAPI::getConfigValue(PHANDLE, "plugin:hyprexpo:skip_empty")->getDataStaticPtr();
-
-    SIDE_LENGTH = **PCOLUMNS;
-    GAP_WIDTH   = **PGAPS;
-    BG_COLOR    = **PCOL;
-
-    // Get workspace method for this specific monitor
-    auto [methodCenter, methodStartID] = getWorkspaceMethodForMonitor(pMonitor.lock());
-
-    images.resize(SIDE_LENGTH * SIDE_LENGTH);
-
-    // r includes empty workspaces; m skips over them
-    std::string selector = **PSKIP ? "m" : "r";
-
-    if (methodCenter) {
-        int currentID = methodStartID;
-        int firstID   = currentID;
-
-        int backtracked = 0;
-
-        // Initialize tiles to WORKSPACE_INVALID; cliking one of these results
-        // in changing to "emptynm" (next empty workspace). Tiles with this id
-        // will only remain if skip_empty is on.
-        for (size_t i = 0; i < images.size(); i++) {
-            images[i].workspaceID = WORKSPACE_INVALID;
-        }
-
-        // Scan through workspaces lower than methodStartID until we wrap; count how many
-        for (size_t i = 1; i < images.size() / 2; ++i) {
-            currentID = getWorkspaceIDNameFromString(selector + "-" + std::to_string(i)).id;
-            if (currentID >= firstID)
-                break;
-
-            backtracked++;
-            firstID = currentID;
-        }
-
-        // Scan through workspaces higher than methodStartID. If using "m"
-        // (skip_empty), stop when we wrap, leaving the rest of the workspace
-        // ID's set to WORKSPACE_INVALID
-        for (size_t i = 0; i < (size_t)(SIDE_LENGTH * SIDE_LENGTH); ++i) {
-            auto& image = images[i];
-            if ((int64_t)i - backtracked < 0) {
-                currentID = getWorkspaceIDNameFromString(selector + std::to_string((int64_t)i - backtracked)).id;
-            } else {
-                currentID = getWorkspaceIDNameFromString(selector + "+" + std::to_string((int64_t)i - backtracked)).id;
-                if (i > 0 && currentID <= firstID)
-                    break;
-            }
-            image.workspaceID = currentID;
-        }
-
-    } else {
-        int currentID         = methodStartID;
-        images[0].workspaceID = currentID;
-
-        auto PWORKSPACESTART = g_pCompositor->getWorkspaceByID(currentID);
-        if (!PWORKSPACESTART)
-            PWORKSPACESTART = CWorkspace::create(currentID, pMonitor.lock(), std::to_string(currentID));
-
-        pMonitor->m_activeWorkspace = PWORKSPACESTART;
-
-        // Scan through workspaces higher than methodStartID. If using "m"
-        // (skip_empty), stop when we wrap, leaving the rest of the workspace
-        // ID's set to WORKSPACE_INVALID
-        for (size_t i = 1; i < (size_t)(SIDE_LENGTH * SIDE_LENGTH); ++i) {
-            auto& image = images[i];
-            currentID   = getWorkspaceIDNameFromString(selector + "+" + std::to_string(i)).id;
-            if (currentID <= methodStartID)
-                break;
-            image.workspaceID = currentID;
-        }
-
-        pMonitor->m_activeWorkspace = startedOn;
-    }
-
-    Render::GL::g_pHyprOpenGL->makeEGLCurrent();
-
-    Vector2D tileSize       = pMonitor->m_size / SIDE_LENGTH;
-    Vector2D tileRenderSize = (pMonitor->m_size - Vector2D{GAP_WIDTH * pMonitor->m_scale, GAP_WIDTH * pMonitor->m_scale} * (SIDE_LENGTH - 1)) / SIDE_LENGTH;
-    CBox     monbox{0, 0, tileSize.x * 2, tileSize.y * 2};
-
-    if (!ENABLE_LOWRES)
-        monbox = {{0, 0}, pMonitor->m_pixelSize};
-
-    int          currentid = 0;
-
-    // Temporarily disable monitor rotation during framebuffer capture so
-    // workspace content renders in the logical (portrait) orientation
-    // rather than the physical panel orientation.
-    const auto savedTransform       = pMonitor->m_transform;
-    const auto savedTransformedSize = pMonitor->m_transformedSize;
-    const auto savedPixelSize       = pMonitor->m_pixelSize;
-
-    // Fix for rotated monitors: m_pixelSize contains physical panel dimensions
-    // (landscape), but we need logical portrait dimensions for the framebuffer
-    if (isTransformRotated(savedTransform)) {
-        // Swap monbox dimensions to match logical orientation
-        monbox = {{0, 0}, {monbox.h, monbox.w}};
-
-        // Override monitor state: disable rotation and set all size fields to
-        // portrait dimensions so beginRender sets up the viewport correctly
-        pMonitor->m_transform       = WL_OUTPUT_TRANSFORM_NORMAL;
-        pMonitor->m_pixelSize       = {monbox.w, monbox.h};
-        pMonitor->m_transformedSize = {monbox.w, monbox.h};
-    }
-
-    PHLWORKSPACE openSpecial = PMONITOR->m_activeSpecialWorkspace;
-    if (openSpecial)
-        PMONITOR->m_activeSpecialWorkspace.reset();
-
-    g_pHyprRenderer->m_bBlockSurfaceFeedback = true;
-
-    startedOn->m_visible = false;
-
-    for (size_t i = 0; i < (size_t)(SIDE_LENGTH * SIDE_LENGTH); ++i) {
-        COverview::SWorkspaceImage& image = images[i];
-        if (!image.fb)
-            image.fb = g_pHyprRenderer->createFB("hyprexpo");
-        image.fb->alloc(monbox.w, monbox.h, framebufferFormatWithAlpha(PMONITOR->m_output->state->state().drmFormat));
-
-        CRegion fakeDamage{0, 0, INT16_MAX, INT16_MAX};
-        g_pHyprRenderer->beginRender(PMONITOR, fakeDamage, Render::RENDER_MODE_FULL_FAKE, nullptr, image.fb);
-
-        clearWithColor(CHyprColor{0, 0, 0, 1.0});
-
-        const auto PWORKSPACE = g_pCompositor->getWorkspaceByID(image.workspaceID);
-
-        if (PWORKSPACE == startedOn)
-            currentid = i;
-
-        if (PWORKSPACE) {
-            image.pWorkspace            = PWORKSPACE;
-            PMONITOR->m_activeWorkspace = PWORKSPACE;
-            const auto previewState     = applyWorkspacePreviewState(PWORKSPACE);
-
-            if (PWORKSPACE == startedOn)
-                PMONITOR->m_activeSpecialWorkspace = openSpecial;
-
-            g_pHyprRenderer->renderWorkspace(PMONITOR, PWORKSPACE, Time::steadyNow(), monbox);
-
-            restoreWorkspacePreviewState(PWORKSPACE, previewState);
-
-            if (PWORKSPACE == startedOn)
-                PMONITOR->m_activeSpecialWorkspace.reset();
-        } else
-            g_pHyprRenderer->renderWorkspace(PMONITOR, PWORKSPACE, Time::steadyNow(), monbox);
-
-        image.box = {(i % SIDE_LENGTH) * tileRenderSize.x + (i % SIDE_LENGTH) * GAP_WIDTH, (i / SIDE_LENGTH) * tileRenderSize.y + (i / SIDE_LENGTH) * GAP_WIDTH, tileRenderSize.x,
-                     tileRenderSize.y};
-
-        g_pHyprRenderer->m_renderData.blockScreenShader = true;
-        g_pHyprRenderer->endRender();
-    }
-
-    g_pHyprRenderer->m_bBlockSurfaceFeedback = false;
-
-    // Restore the original monitor state after capture
-    pMonitor->m_transform       = savedTransform;
-    pMonitor->m_pixelSize       = savedPixelSize;
-    pMonitor->m_transformedSize = savedTransformedSize;
-
-    PMONITOR->m_activeSpecialWorkspace = openSpecial;
-    PMONITOR->m_activeWorkspace        = startedOn;
-    startedOn->m_visible               = true;
-    g_pDesktopAnimationManager->startAnimation(startedOn, CDesktopAnimationManager::ANIMATION_TYPE_IN, true, true);
-
-    // zoom on the current workspace.
-    // const auto& TILE = images[std::clamp(currentid, 0, SIDE_LENGTH * SIDE_LENGTH)];
-
-    g_pAnimationManager->createAnimation(pMonitor->m_size * pMonitor->m_size / tileSize, size, Config::animationTree()->getAnimationPropertyConfig("windowsMove"), AVARDAMAGE_NONE);
-    g_pAnimationManager->createAnimation((-((pMonitor->m_size / (double)SIDE_LENGTH) * Vector2D{currentid % SIDE_LENGTH, currentid / SIDE_LENGTH}) * pMonitor->m_scale) *
-                                             (pMonitor->m_size / tileSize),
-                                         pos, Config::animationTree()->getAnimationPropertyConfig("windowsMove"), AVARDAMAGE_NONE);
-
-    size->setUpdateCallback(damageMonitor);
-    pos->setUpdateCallback(damageMonitor);
-
-    if (!swipe) {
-        *size = pMonitor->m_size;
-        *pos  = {0, 0};
-
-        size->setCallbackOnEnd([this](auto) { redrawAll(true); });
-    }
-
-    openedID = currentid;
-
-    ensureOverviewCursorVisible(true);
-
-    lastMousePosLocal = g_pInputManager->getMouseCoordsInternal() - pMonitor->m_position;
-
-    // Initialize hoveredID based on current mouse position
-    int hx = std::clamp((int)(lastMousePosLocal.x / pMonitor->m_size.x * SIDE_LENGTH), 0, SIDE_LENGTH - 1);
-    int hy = std::clamp((int)(lastMousePosLocal.y / pMonitor->m_size.y * SIDE_LENGTH), 0, SIDE_LENGTH - 1);
-    hoveredID = hx + hy * SIDE_LENGTH;
-
-    mouseMoveHook = Event::bus()->m_events.input.mouse.move.listen([this](const Vector2D& coords, Event::SCallbackInfo& info) {
-        if (closing)
-            return;
-
-        ensureOverviewCursorVisible(false);
-
-        info.cancelled    = true;
-        lastMousePosLocal = g_pInputManager->getMouseCoordsInternal() - pMonitor->m_position;
-
-        // Update hovered tile
-        int hx = std::clamp((int)(lastMousePosLocal.x / pMonitor->m_size.x * SIDE_LENGTH), 0, SIDE_LENGTH - 1);
-        int hy = std::clamp((int)(lastMousePosLocal.y / pMonitor->m_size.y * SIDE_LENGTH), 0, SIDE_LENGTH - 1);
-        int newHoveredID = hx + hy * SIDE_LENGTH;
-
-        if (newHoveredID != hoveredID) {
-            hoveredID = newHoveredID;
-            damage();
-        }
-    });
-
-    touchMoveHook = Event::bus()->m_events.input.touch.motion.listen([this](const ITouch::SMotionEvent& e, Event::SCallbackInfo& info) {
-        if (closing)
-            return;
-
-        info.cancelled    = true;
-        lastMousePosLocal = g_pInputManager->getMouseCoordsInternal() - pMonitor->m_position;
-
-        int hx = std::clamp((int)(lastMousePosLocal.x / pMonitor->m_size.x * SIDE_LENGTH), 0, SIDE_LENGTH - 1);
-        int hy = std::clamp((int)(lastMousePosLocal.y / pMonitor->m_size.y * SIDE_LENGTH), 0, SIDE_LENGTH - 1);
-        int newHoveredID = hx + hy * SIDE_LENGTH;
-
-        if (newHoveredID != hoveredID) {
-            hoveredID = newHoveredID;
-            damage();
-        }
-    });
-
-    mouseButtonHook = Event::bus()->m_events.input.mouse.button.listen([this](const IPointer::SButtonEvent& e, Event::SCallbackInfo& info) {
-        if (closing)
-            return;
-
-        // If expo hasn't animated in enough to be visible, close silently without
-        // consuming the event. This prevents phantom triggers (e.g. a bouncing
-        // mouse side-button firing the expo keybind) from swallowing real clicks.
-        if (size->getPercent() < 0.05f) {
-            close();
-            return;
-        }
-
-        info.cancelled = true;
-
-        selectHoveredWorkspace();
-
-        close();
-    });
-
-    touchDownHook = Event::bus()->m_events.input.touch.down.listen([this](const ITouch::SDownEvent& e, Event::SCallbackInfo& info) {
-        if (closing)
-            return;
-
-        info.cancelled = true;
-
-        selectHoveredWorkspace();
-
-        close();
-    });
-
-    enterSubmapIfEnabled();
-}
-
-void COverview::selectHoveredWorkspace() {
-    if (closing)
-        return;
-
-    // get tile x,y
-    int x     = std::clamp((int)(lastMousePosLocal.x / pMonitor->m_size.x * SIDE_LENGTH), 0, SIDE_LENGTH - 1);
-    int y     = std::clamp((int)(lastMousePosLocal.y / pMonitor->m_size.y * SIDE_LENGTH), 0, SIDE_LENGTH - 1);
-    closeOnID = x + y * SIDE_LENGTH;
-}
-
-void COverview::ensureKbFocusInitialized() {
-    if (kbFocusID != -1)
-        return;
-
-    // try to set to current openedID
-    if (openedID != -1) {
-        kbFocusID = openedID;
-        return;
-    }
-
-    // fallback: first valid tile
-    for (size_t i = 0; i < images.size(); ++i) {
-        if (isTileValid(i)) {
-            kbFocusID = i;
-            return;
-        }
-    }
-}
-
-bool COverview::isTileValid(int id) const {
-    if (id < 0 || id >= SIDE_LENGTH * SIDE_LENGTH)
-        return false;
-    return images[id].workspaceID != WORKSPACE_INVALID;
-}
-
-int COverview::tileForWorkspaceID(int wsid) const {
-    for (size_t i = 0; i < images.size(); ++i) {
-        if (images[i].workspaceID == wsid)
-            return (int)i;
-    }
-    return -1;
-}
-
-int COverview::tileForVisibleIndex(int vIdx) const {
-    if (vIdx < 0)
-        return -1;
-    int seen = 0;
-    for (size_t i = 0; i < images.size(); ++i) {
-        if (images[i].workspaceID == WORKSPACE_INVALID)
-            continue;
-        if (seen == vIdx)
-            return (int)i;
-        ++seen;
-    }
-    return -1;
-}
-
-bool COverview::selectVisibleToken(const std::string& token) {
-    if (closing)
-        return false;
-
-    static auto* const* PSELECTLABEL = (Hyprlang::INT* const*)HyprlandAPI::getConfigValue(PHANDLE, "plugin:hyprexpo:selection_label_enable")->getDataStaticPtr();
-    static auto const*  PSELECTMAP   = (Hyprlang::STRING const*)HyprlandAPI::getConfigValue(PHANDLE, "plugin:hyprexpo:selection_label_token_map")->getDataStaticPtr();
-
-    const std::string normalized = lowerString(trimString(token));
-    if (normalized.empty())
-        return false;
-
-    if (**PSELECTLABEL) {
-        const auto tokens = splitCommaList(std::string{*PSELECTMAP});
-        for (size_t i = 0; i < tokens.size(); ++i) {
-            if (tokens[i].empty() || lowerString(tokens[i]) != normalized)
-                continue;
-
-            const int tid = tileForVisibleIndex(i);
-            if (tid == -1)
-                return false;
-
-            closeOnID = tid;
-            return true;
-        }
-
-        return false;
-    }
-
-    const int visibleIndex = fallbackTokenToVisibleIndex(normalized);
-    if (visibleIndex < 0)
-        return false;
-
-    const int tid = tileForVisibleIndex(visibleIndex);
-    if (tid == -1)
-        return false;
-
-    closeOnID = tid;
-    return true;
-}
-
-void COverview::moveFocus(int dx, int dy) {
-    ensureKbFocusInitialized();
-    if (kbFocusID == -1)
-        return;
-
-    int x = kbFocusID % SIDE_LENGTH;
-    int y = kbFocusID / SIDE_LENGTH;
-
-    static auto* const* PWRAPH = (Hyprlang::INT* const*)HyprlandAPI::getConfigValue(PHANDLE, "plugin:hyprexpo:keynav_wrap_h")->getDataStaticPtr();
-    static auto* const* PWRAPV = (Hyprlang::INT* const*)HyprlandAPI::getConfigValue(PHANDLE, "plugin:hyprexpo:keynav_wrap_v")->getDataStaticPtr();
-
-    if (dx != 0) {
-        static auto* const* PREADING = (Hyprlang::INT* const*)HyprlandAPI::getConfigValue(PHANDLE, "plugin:hyprexpo:keynav_reading_order")->getDataStaticPtr();
-        int                 step     = dx > 0 ? 1 : -1;
-        if (**PREADING) {
-            // reading-order scan: proceed linearly across the grid (row-major)
-            const int total = SIDE_LENGTH * SIDE_LENGTH;
-            int       idx   = kbFocusID;
-            for (int tries = 0; tries < total; ++tries) {
-                idx += step;
-                if (idx < 0 || idx >= total) {
-                    // wrap only if both wraps are enabled (edge of grid)
-                    if (**PWRAPH && **PWRAPV)
-                        idx = (idx + total) % total;
-                    else
-                        break;
-                }
-                if (isTileValid(idx)) {
-                    kbFocusID = idx;
-                    return;
-                }
-            }
-        } else {
-            // in-row scan with optional horizontal wrap
-            int nx = x;
-            for (int tries = 0; tries < SIDE_LENGTH; ++tries) {
-                nx += step;
-                if (nx < 0 || nx >= SIDE_LENGTH) {
-                    if (**PWRAPH)
-                        nx = (nx + SIDE_LENGTH) % SIDE_LENGTH;
-                    else
-                        break;
-                }
-                const int nid = nx + y * SIDE_LENGTH;
-                if (isTileValid(nid)) {
-                    kbFocusID = nid;
-                    return;
-                }
-            }
-        }
-    }
-
-    if (dy != 0) {
-        int step = dy > 0 ? 1 : -1;
-        int ny   = y;
-        for (int tries = 0; tries < SIDE_LENGTH; ++tries) {
-            ny += step;
-            if (ny < 0 || ny >= SIDE_LENGTH) {
-                if (**PWRAPV)
-                    ny = (ny + SIDE_LENGTH) % SIDE_LENGTH;
-                else
-                    break;
-            }
-            const int nid = x + ny * SIDE_LENGTH;
-            if (isTileValid(nid)) {
-                kbFocusID = nid;
-                return;
-            }
-        }
-    }
-}
-
-void COverview::onKbMoveFocus(const std::string& dir) {
-    if (closing)
-        return;
-    if (dir == "left")
-        moveFocus(-1, 0);
-    else if (dir == "right")
-        moveFocus(1, 0);
-    else if (dir == "up")
-        moveFocus(0, -1);
-    else if (dir == "down")
-        moveFocus(0, 1);
-
-    damage();
-}
-
-void COverview::onKbConfirm() {
-    if (closing)
-        return;
-    ensureKbFocusInitialized();
-    if (kbFocusID != -1)
-        closeOnID = kbFocusID;
-    close();
-}
-
-void COverview::onKbSelectNumber(int num) {
-    if (closing)
-        return;
-
-    if (num == 0)
-        num = 10;
-
-    const int tid = tileForWorkspaceID(num);
-    if (tid != -1) {
-        closeOnID = tid;
-        close();
-    }
-}
-
-void COverview::onKbSelectToken(int visibleIdx) {
-    if (closing)
-        return;
-    if (visibleIdx < 0)
-        return;
-    const int tid = tileForVisibleIndex(visibleIdx);
-    if (tid != -1) {
-        closeOnID = tid;
-        close();
-    }
-}
-
-void COverview::redrawID(int id, bool forcelowres) {
-    if (pMonitor->m_activeWorkspace != startedOn && !closing) {
-        // likely user changed.
-        onWorkspaceChange();
-    }
-
-    blockOverviewRendering = true;
-
-    Render::GL::g_pHyprOpenGL->makeEGLCurrent();
-
-    id = std::clamp(id, 0, SIDE_LENGTH * SIDE_LENGTH - 1);
-
-    Vector2D tileSize       = pMonitor->m_size / SIDE_LENGTH;
-    Vector2D tileRenderSize = (pMonitor->m_size - Vector2D{GAP_WIDTH, GAP_WIDTH} * (SIDE_LENGTH - 1)) / SIDE_LENGTH;
-    CBox     monbox{0, 0, tileSize.x * 2, tileSize.y * 2};
-
-    if (!forcelowres && (size->value() != pMonitor->m_size || closing))
-        monbox = {{0, 0}, pMonitor->m_pixelSize};
-
-    if (!ENABLE_LOWRES)
-        monbox = {{0, 0}, pMonitor->m_pixelSize};
-
-    const auto savedTransform       = pMonitor->m_transform;
-    const auto savedTransformedSize = pMonitor->m_transformedSize;
-    const auto savedPixelSize       = pMonitor->m_pixelSize;
-
-    // Fix for rotated monitors: swap dimensions to match logical orientation
-    if (isTransformRotated(savedTransform)) {
-        monbox = {{0, 0}, {monbox.h, monbox.w}};
-
-        // Override monitor state to disable rotation
-        pMonitor->m_transform       = WL_OUTPUT_TRANSFORM_NORMAL;
-        pMonitor->m_pixelSize       = {monbox.w, monbox.h};
-        pMonitor->m_transformedSize = {monbox.w, monbox.h};
-    }
-
-    auto& image = images[id];
-
-    if (!image.fb)
-        image.fb = g_pHyprRenderer->createFB("hyprexpo");
-    if (image.fb->m_size != monbox.size()) {
-        image.fb->release();
-        image.fb->alloc(monbox.w, monbox.h, framebufferFormatWithAlpha(pMonitor->m_output->state->state().drmFormat));
-    }
-
-    CRegion fakeDamage{0, 0, INT16_MAX, INT16_MAX};
-    g_pHyprRenderer->beginRender(pMonitor.lock(), fakeDamage, Render::RENDER_MODE_FULL_FAKE, nullptr, image.fb);
-
-    clearWithColor(CHyprColor{0, 0, 0, 1.0});
-
-    const auto   PWORKSPACE = image.pWorkspace;
-
-    PHLWORKSPACE openSpecial = pMonitor->m_activeSpecialWorkspace;
-    if (openSpecial)
-        pMonitor->m_activeSpecialWorkspace.reset();
-
-    startedOn->m_visible = false;
-
-    if (PWORKSPACE) {
-        pMonitor->m_activeWorkspace = PWORKSPACE;
-        const auto previewState     = applyWorkspacePreviewState(PWORKSPACE);
-
-        if (PWORKSPACE == startedOn)
-            pMonitor->m_activeSpecialWorkspace = openSpecial;
-
-        g_pHyprRenderer->renderWorkspace(pMonitor.lock(), PWORKSPACE, Time::steadyNow(), monbox);
-
-        restoreWorkspacePreviewState(PWORKSPACE, previewState);
-
-        if (PWORKSPACE == startedOn)
-            pMonitor->m_activeSpecialWorkspace.reset();
-    } else
-        g_pHyprRenderer->renderWorkspace(pMonitor.lock(), PWORKSPACE, Time::steadyNow(), monbox);
-
-    g_pHyprRenderer->m_renderData.blockScreenShader = true;
-    g_pHyprRenderer->endRender();
-
-    // Restore the original monitor state after capture
-    pMonitor->m_transform       = savedTransform;
-    pMonitor->m_pixelSize       = savedPixelSize;
-    pMonitor->m_transformedSize = savedTransformedSize;
-
-    pMonitor->m_activeSpecialWorkspace = openSpecial;
-    pMonitor->m_activeWorkspace        = startedOn;
-    startedOn->m_visible               = true;
-    g_pDesktopAnimationManager->startAnimation(startedOn, CDesktopAnimationManager::ANIMATION_TYPE_IN, true, true);
-
-    blockOverviewRendering = false;
-}
-
-void COverview::redrawAll(bool forcelowres) {
-    for (size_t i = 0; i < (size_t)(SIDE_LENGTH * SIDE_LENGTH); ++i) {
-        redrawID(i, forcelowres);
-    }
-}
-
-void COverview::damage() {
-    blockDamageReporting = true;
-    g_pHyprRenderer->damageMonitor(pMonitor.lock());
-    blockDamageReporting = false;
-}
-
-void COverview::onDamageReported() {
-    damageDirty = true;
-
-    Vector2D SIZE = size->value();
-
-    Vector2D tileSize       = (SIZE / SIDE_LENGTH);
-    const auto GAPSIZE      = (closing ? (1.0 - size->getPercent()) : size->getPercent()) * GAP_WIDTH;
-    static auto* const* PGAPSO = (Hyprlang::INT* const*)HyprlandAPI::getConfigValue(PHANDLE, "plugin:hyprexpo:gaps_out")->getDataStaticPtr();
-    const float OUTER       = **PGAPSO * (closing ? (1.0 - size->getPercent()) : size->getPercent());
-    Vector2D tileRenderSize = (SIZE - Vector2D{GAPSIZE, GAPSIZE} * (SIDE_LENGTH - 1) - Vector2D{OUTER * 2, OUTER * 2}) / SIDE_LENGTH;
-    // const auto& TILE           = images[std::clamp(openedID, 0, SIDE_LENGTH * SIDE_LENGTH)];
-    CBox texbox = CBox{OUTER + (openedID % SIDE_LENGTH) * tileRenderSize.x + (openedID % SIDE_LENGTH) * GAPSIZE,
-                       OUTER + (openedID / SIDE_LENGTH) * tileRenderSize.y + (openedID / SIDE_LENGTH) * GAPSIZE, tileRenderSize.x, tileRenderSize.y}
-                      .translate(pMonitor->m_position);
-
-    damage();
-
-    blockDamageReporting = true;
-    g_pHyprRenderer->damageBox(texbox);
-    blockDamageReporting = false;
-    g_pCompositor->scheduleFrameForMonitor(pMonitor.lock());
-}
-
-void COverview::close(bool switchToSelection) {
-    if (closing)
-        return;
-
-    resetSubmapIfNeeded();
-
-    const int   ID = closeOnID == -1 ? openedID : closeOnID;
-
-    const int   SAFEID = std::clamp(ID, 0, SIDE_LENGTH * SIDE_LENGTH - 1);
-    const auto& TILE   = images[SAFEID];
-
-    Vector2D    tileSize = (pMonitor->m_size / SIDE_LENGTH);
-
-    *size = pMonitor->m_size * pMonitor->m_size / tileSize;
-    *pos  = (-((pMonitor->m_size / (double)SIDE_LENGTH) * Vector2D{SAFEID % SIDE_LENGTH, SAFEID / SIDE_LENGTH}) * pMonitor->m_scale) * (pMonitor->m_size / tileSize);
-
-    size->setCallbackOnEnd(removeOverview);
-
-    closing = true;
-
-    redrawAll();
-
-    if (switchToSelection && TILE.workspaceID != pMonitor->activeWorkspaceID()) {
-        pMonitor->setSpecialWorkspace(0);
-
-        // If this tile's workspace was WORKSPACE_INVALID, move to the next
-        // empty workspace. This should only happen if skip_empty is on, in
-        // which case some tiles will be left with this ID intentionally.
-        const int  NEWID = TILE.workspaceID == WORKSPACE_INVALID ? getWorkspaceIDNameFromString("emptynm").id : TILE.workspaceID;
-
-        const auto NEWIDWS = g_pCompositor->getWorkspaceByID(NEWID);
-
-        const auto OLDWS = pMonitor->m_activeWorkspace;
-
-        const auto CHANGE = !NEWIDWS ? Config::Actions::changeWorkspace(std::to_string(NEWID)) : Config::Actions::changeWorkspace(NEWIDWS->getConfigName());
-        if (!CHANGE)
-            Log::logger->log(Log::ERR, "[hyprexpo] failed to change workspace: {}", CHANGE.error().message);
-
-        g_pDesktopAnimationManager->startAnimation(pMonitor->m_activeWorkspace, CDesktopAnimationManager::ANIMATION_TYPE_IN, true, true);
-        g_pDesktopAnimationManager->startAnimation(OLDWS, CDesktopAnimationManager::ANIMATION_TYPE_OUT, false, true);
-
-        startedOn = pMonitor->m_activeWorkspace;
-    }
-}
-
-void COverview::onPreRender() {
-    if (damageDirty) {
-        damageDirty = false;
-        redrawID(closing ? (closeOnID == -1 ? openedID : closeOnID) : openedID);
-    }
-}
-
-void COverview::onWorkspaceChange() {
-    if (valid(startedOn))
-        g_pDesktopAnimationManager->startAnimation(startedOn, CDesktopAnimationManager::ANIMATION_TYPE_OUT, false, true);
-    else
-        startedOn = pMonitor->m_activeWorkspace;
-
-    for (size_t i = 0; i < (size_t)(SIDE_LENGTH * SIDE_LENGTH); ++i) {
-        if (images[i].workspaceID != pMonitor->activeWorkspaceID())
-            continue;
-
-        openedID = i;
-        break;
-    }
-
-    closeOnID = openedID;
-    close();
-}
-
-void COverview::render() {
-    g_pHyprRenderer->m_renderPass.add(makeUnique<COverviewPassElement>());
-}
-
-void COverview::fullRender() {
-    const auto GAPSIZE = (closing ? (1.0 - size->getPercent()) : size->getPercent()) * GAP_WIDTH;
-
-    if (pMonitor->m_activeWorkspace != startedOn && !closing) {
-        // likely user changed.
-        onWorkspaceChange();
-    }
-
-    Vector2D SIZE = size->value();
-
-    static auto* const* PGAPSO = (Hyprlang::INT* const*)HyprlandAPI::getConfigValue(PHANDLE, "plugin:hyprexpo:gaps_out")->getDataStaticPtr();
-    const float OUTER = **PGAPSO * (closing ? (1.0 - size->getPercent()) : size->getPercent());
-
-    Vector2D tileSize       = (SIZE / SIDE_LENGTH);
-    Vector2D tileRenderSize = (SIZE - Vector2D{GAPSIZE, GAPSIZE} * (SIDE_LENGTH - 1) - Vector2D{OUTER * 2, OUTER * 2}) / SIDE_LENGTH;
-
-    clearWithColor(BG_COLOR.stripA());
-
-    static auto* const* PTILEROUND = (Hyprlang::INT* const*)HyprlandAPI::getConfigValue(PHANDLE, "plugin:hyprexpo:tile_rounding")->getDataStaticPtr();
-    static auto* const* PTOUNDPWR  = (Hyprlang::FLOAT* const*)HyprlandAPI::getConfigValue(PHANDLE, "plugin:hyprexpo:tile_rounding_power")->getDataStaticPtr();
-    static auto* const* PTILEROUNDF = (Hyprlang::INT* const*)HyprlandAPI::getConfigValue(PHANDLE, "plugin:hyprexpo:tile_rounding_focus")->getDataStaticPtr();
-    static auto* const* PTILEROUNDC = (Hyprlang::INT* const*)HyprlandAPI::getConfigValue(PHANDLE, "plugin:hyprexpo:tile_rounding_current")->getDataStaticPtr();
-    static auto* const* PTILEROUNDH = (Hyprlang::INT* const*)HyprlandAPI::getConfigValue(PHANDLE, "plugin:hyprexpo:tile_rounding_hover")->getDataStaticPtr();
-
-    const int BASE_ROUND_SCALED   = std::max(0, (int)std::lround((double)**PTILEROUND * pMonitor->m_scale));
-    const int FOCUS_ROUND_SCALED  = **PTILEROUNDF >= 0 ? std::max(0, (int)std::lround((double)**PTILEROUNDF * pMonitor->m_scale)) : BASE_ROUND_SCALED;
-    const int CURRENT_ROUND_SCALED= **PTILEROUNDC >= 0 ? std::max(0, (int)std::lround((double)**PTILEROUNDC * pMonitor->m_scale)) : BASE_ROUND_SCALED;
-    const int HOVER_ROUND_SCALED  = **PTILEROUNDH >= 0 ? std::max(0, (int)std::lround((double)**PTILEROUNDH * pMonitor->m_scale)) : BASE_ROUND_SCALED;
-    const float ROUND_PWR         = **PTOUNDPWR;
-
-    // (shadows moved to feature/shadows branch)
-
-    for (size_t y = 0; y < (size_t)SIDE_LENGTH; ++y) {
-        for (size_t x = 0; x < (size_t)SIDE_LENGTH; ++x) {
-            const int id = x + y * SIDE_LENGTH;
-            CBox      texbox{OUTER + x * tileRenderSize.x + x * GAPSIZE, OUTER + y * tileRenderSize.y + y * GAPSIZE, tileRenderSize.x, tileRenderSize.y};
-            texbox.scale(pMonitor->m_scale).translate(pos->value());
-            texbox.round();
-            // per-tile rounding override for focus/current/hover (priority: focus > current > hover)
-            int tileRound = BASE_ROUND_SCALED;
-            if ((int)id == kbFocusID)
-                tileRound = FOCUS_ROUND_SCALED;
-            else if ((int)id == openedID)
-                tileRound = CURRENT_ROUND_SCALED;
-            else if ((int)id == hoveredID)
-                tileRound = HOVER_ROUND_SCALED;
-
-            // clamp rounding to tile size
-            const int maxCornerPx = std::max(0, (int)std::floor(std::min(texbox.w, texbox.h) / 2.0));
-            tileRound = std::min(tileRound, maxCornerPx);
-
-            // no shadow in this branch
-
-            CRegion damage{0, 0, INT16_MAX, INT16_MAX};
-            Render::GL::g_pHyprOpenGL->renderTextureInternal(images[id].fb->getTexture(), texbox, {.damage = &damage, .a = 1.0, .round = tileRound, .roundingPower = ROUND_PWR});
-        }
-    }
-
-    // overlays: numbers and borders
-    static auto* const* PLABELEN   = (Hyprlang::INT* const*)HyprlandAPI::getConfigValue(PHANDLE, "plugin:hyprexpo:label_enable")->getDataStaticPtr();
-    static auto* const* PLABELSIZE = (Hyprlang::INT* const*)HyprlandAPI::getConfigValue(PHANDLE, "plugin:hyprexpo:label_font_size")->getDataStaticPtr();
-    static auto  const* PLABELPOS  = (Hyprlang::STRING const*)HyprlandAPI::getConfigValue(PHANDLE, "plugin:hyprexpo:label_position")->getDataStaticPtr();
-    static auto  const* PLABELMODE = (Hyprlang::STRING const*)HyprlandAPI::getConfigValue(PHANDLE, "plugin:hyprexpo:label_text_mode")->getDataStaticPtr();
-    static auto  const* PTOKENMAP  = (Hyprlang::STRING const*)HyprlandAPI::getConfigValue(PHANDLE, "plugin:hyprexpo:label_token_map")->getDataStaticPtr();
-    static auto* const* PLABELOX   = (Hyprlang::INT* const*)HyprlandAPI::getConfigValue(PHANDLE, "plugin:hyprexpo:label_offset_x")->getDataStaticPtr();
-    static auto* const* PLABELOY   = (Hyprlang::INT* const*)HyprlandAPI::getConfigValue(PHANDLE, "plugin:hyprexpo:label_offset_y")->getDataStaticPtr();
-    static auto  const* PLABELSHOW = (Hyprlang::STRING const*)HyprlandAPI::getConfigValue(PHANDLE, "plugin:hyprexpo:label_show")->getDataStaticPtr();
-    static auto* const* PLCOLDEF   = (Hyprlang::INT* const*)HyprlandAPI::getConfigValue(PHANDLE, "plugin:hyprexpo:label_color_default")->getDataStaticPtr();
-    static auto* const* PLCOLHOV   = (Hyprlang::INT* const*)HyprlandAPI::getConfigValue(PHANDLE, "plugin:hyprexpo:label_color_hover")->getDataStaticPtr();
-    static auto* const* PLCOLFOC   = (Hyprlang::INT* const*)HyprlandAPI::getConfigValue(PHANDLE, "plugin:hyprexpo:label_color_focus")->getDataStaticPtr();
-    static auto* const* PLCOLCUR   = (Hyprlang::INT* const*)HyprlandAPI::getConfigValue(PHANDLE, "plugin:hyprexpo:label_color_current")->getDataStaticPtr();
-    static auto* const* PLSCALEH   = (Hyprlang::FLOAT* const*)HyprlandAPI::getConfigValue(PHANDLE, "plugin:hyprexpo:label_scale_hover")->getDataStaticPtr();
-    static auto* const* PLSCALEF   = (Hyprlang::FLOAT* const*)HyprlandAPI::getConfigValue(PHANDLE, "plugin:hyprexpo:label_scale_focus")->getDataStaticPtr();
-    static auto* const* PLBGEN     = (Hyprlang::INT* const*)HyprlandAPI::getConfigValue(PHANDLE, "plugin:hyprexpo:label_bg_enable")->getDataStaticPtr();
-    static auto* const* PLBGCOL    = (Hyprlang::INT* const*)HyprlandAPI::getConfigValue(PHANDLE, "plugin:hyprexpo:label_bg_color")->getDataStaticPtr();
-    static auto* const* PLBGROUND  = (Hyprlang::INT* const*)HyprlandAPI::getConfigValue(PHANDLE, "plugin:hyprexpo:label_bg_rounding")->getDataStaticPtr();
-    static auto  const* PLBGSHAPE  = (Hyprlang::STRING const*)HyprlandAPI::getConfigValue(PHANDLE, "plugin:hyprexpo:label_bg_shape")->getDataStaticPtr();
-    static auto* const* PLBGPAD    = (Hyprlang::INT* const*)HyprlandAPI::getConfigValue(PHANDLE, "plugin:hyprexpo:label_padding")->getDataStaticPtr();
-
-    static auto* const* PBWIDTH      = (Hyprlang::INT* const*)HyprlandAPI::getConfigValue(PHANDLE, "plugin:hyprexpo:border_width")->getDataStaticPtr();
-    static auto  const* PBCOLCUR     = (Hyprlang::STRING const*)HyprlandAPI::getConfigValue(PHANDLE, "plugin:hyprexpo:border_color_current")->getDataStaticPtr();
-    static auto  const* PBCOLFOC     = (Hyprlang::STRING const*)HyprlandAPI::getConfigValue(PHANDLE, "plugin:hyprexpo:border_color_focus")->getDataStaticPtr();
-    static auto  const* PBCOLHOV     = (Hyprlang::STRING const*)HyprlandAPI::getConfigValue(PHANDLE, "plugin:hyprexpo:border_color_hover")->getDataStaticPtr();
-    // Deprecated configs for backwards compatibility
-    static auto  const* PBGRCUR      = (Hyprlang::STRING const*)HyprlandAPI::getConfigValue(PHANDLE, "plugin:hyprexpo:border_grad_current")->getDataStaticPtr();
-    static auto  const* PBGREFOC     = (Hyprlang::STRING const*)HyprlandAPI::getConfigValue(PHANDLE, "plugin:hyprexpo:border_grad_focus")->getDataStaticPtr();
-    static auto  const* PBGREHOV     = (Hyprlang::STRING const*)HyprlandAPI::getConfigValue(PHANDLE, "plugin:hyprexpo:border_grad_hover")->getDataStaticPtr();
-
-    static auto* const* PSELECTEN   = (Hyprlang::INT* const*)HyprlandAPI::getConfigValue(PHANDLE, "plugin:hyprexpo:selection_label_enable")->getDataStaticPtr();
-    static auto  const* PSELECTMAP  = (Hyprlang::STRING const*)HyprlandAPI::getConfigValue(PHANDLE, "plugin:hyprexpo:selection_label_token_map")->getDataStaticPtr();
-    static auto  const* PSELECTPOS  = (Hyprlang::STRING const*)HyprlandAPI::getConfigValue(PHANDLE, "plugin:hyprexpo:selection_label_position")->getDataStaticPtr();
-    static auto* const* PSELECTOX   = (Hyprlang::INT* const*)HyprlandAPI::getConfigValue(PHANDLE, "plugin:hyprexpo:selection_label_offset_x")->getDataStaticPtr();
-    static auto* const* PSELECTOY   = (Hyprlang::INT* const*)HyprlandAPI::getConfigValue(PHANDLE, "plugin:hyprexpo:selection_label_offset_y")->getDataStaticPtr();
-    static auto* const* PSELECTCOL  = (Hyprlang::INT* const*)HyprlandAPI::getConfigValue(PHANDLE, "plugin:hyprexpo:selection_label_color")->getDataStaticPtr();
-
-    // draw labels
-    if (**PLABELEN || **PSELECTEN) {
-        // use the tracked hoveredID (cleared during closing)
-        const int labelHoveredID = closing ? -1 : hoveredID;
-
-        auto shouldShow = [&](int id) -> bool {
-            if (std::string{*PLABELSHOW} == "never")
-                return false;
-            if (std::string{*PLABELSHOW} == "always")
-                return true;
-            const bool isHover  = id == labelHoveredID;
-            const bool isFocus  = id == kbFocusID;
-            const bool isCurr   = id == openedID;
-            const std::string mode{*PLABELSHOW};
-            if (mode == "hover")
-                return isHover;
-            if (mode == "focus")
-                return isFocus;
-            if (mode == "hover+focus")
-                return isHover || isFocus;
-            if (mode == "current+focus")
-                return isCurr || isFocus;
-            return true;
-        };
-
-        auto resolveState = [&](int id) -> int {
-            // precedence: focus > current > hover > default
-            if (id == kbFocusID)
-                return 2; // focus
-            if (id == openedID)
-                return 3; // current
-            if (id == labelHoveredID)
-                return 1; // hover
-            return 0;     // default
-        };
-
-        auto placeBox = [&](const CBox& tile, const Vector2D& size, const std::string& anchor, int offsetX, int offsetY) -> CBox {
-            double x = tile.x, y = tile.y;
-            if (anchor == "top-left") {
-                x += offsetX; y += offsetY;
-            } else if (anchor == "top-right") {
-                x += tile.w - size.x - offsetX; y += offsetY;
-            } else if (anchor == "bottom-left") {
-                x += offsetX; y += tile.h - size.y - offsetY;
-            } else if (anchor == "bottom-right") {
-                x += tile.w - size.x - offsetX; y += tile.h - size.y - offsetY;
-            } else { // center
-                x += (tile.w - size.x) / 2.0; y += (tile.h - size.y) / 2.0;
-            }
-            return CBox{x, y, (double)size.x, (double)size.y};
-        };
-
-        auto renderLabel = [&](SP<Render::ITexture>& tex, Vector2D& sz, const std::string& label, const CHyprColor& col, float scaleMul, const CBox& tile, const std::string& anchor,
-                               int offsetX, int offsetY) {
-            if (label.empty())
-                return;
-
-            const int baseF = std::max(8, (int)**PLABELSIZE);
-            if (!tex || tex->m_texID == 0) {
-                const int fsz = std::max(8, (int)std::round(baseF * scaleMul));
-                Vector2D  buf{std::max(32, fsz * 2), std::max(24, fsz + 8)};
-                sz  = buf;
-                tex = renderNumberTexture(label, col, buf, pMonitor->m_scale, fsz);
-            }
-
-            if (!tex || tex->m_texID == 0)
-                return;
-
-            static auto* const* PLPIXELSNAP = (Hyprlang::INT* const*)HyprlandAPI::getConfigValue(PHANDLE, "plugin:hyprexpo:label_pixel_snap")->getDataStaticPtr();
-
-            auto drawWithBG = [&]() {
-                const int pad = **PLBGPAD;
-                Vector2D  bgSize = {sz.x + pad * 2, sz.y + pad * 2};
-                const std::string shape{*PLBGSHAPE};
-                int roundPx = **PLBGROUND;
-                if (shape == "circle" || shape == "square") {
-                    const double side = std::max(bgSize.x, bgSize.y);
-                    bgSize            = {side, side};
-                    roundPx           = (shape == "circle") ? std::lround(side / 2.0) : 0;
-                }
-                CBox bg = placeBox(tile, bgSize, anchor, offsetX, offsetY);
-                CBox lb{bg.x + (bg.w - sz.x) / 2.0, bg.y + (bg.h - sz.y) / 2.0, (double)sz.x, (double)sz.y};
-                if (**PLPIXELSNAP) {
-                    bg.round();
-                    lb.round();
-                }
-                Render::GL::g_pHyprOpenGL->renderRect(bg, CHyprColor{(uint64_t)**PLBGCOL}, {.round = roundPx});
-                Render::GL::g_pHyprOpenGL->renderTexture(tex, lb, {.a = 1.0});
-            };
-
-            auto drawNoBG = [&]() {
-                CBox lb = placeBox(tile, sz, anchor, offsetX, offsetY);
-                if (**PLPIXELSNAP)
-                    lb.round();
-                Render::GL::g_pHyprOpenGL->renderTexture(tex, lb, {.a = 1.0});
-            };
-
-            if (**PLBGEN)
-                drawWithBG();
-            else
-                drawNoBG();
-        };
-
-        std::vector<std::string> labelTokens;
-        if (!std::string{*PTOKENMAP}.empty())
-            labelTokens = splitCommaList(std::string{*PTOKENMAP});
-
-        std::vector<std::string> selectionTokens;
-        if (!std::string{*PSELECTMAP}.empty())
-            selectionTokens = splitCommaList(std::string{*PSELECTMAP});
-
-        int tokenCounter = 0;
-        for (size_t y = 0; y < (size_t)SIDE_LENGTH; ++y) {
-            for (size_t x = 0; x < (size_t)SIDE_LENGTH; ++x) {
-                const int id = x + y * SIDE_LENGTH;
-                if (images[id].workspaceID == WORKSPACE_INVALID)
-                    continue;
-
-                // compute tile box again for label placement
-                CBox tile{OUTER + x * tileRenderSize.x + x * GAPSIZE, OUTER + y * tileRenderSize.y + y * GAPSIZE, tileRenderSize.x, tileRenderSize.y};
-                tile.scale(pMonitor->m_scale).translate(pos->value());
-                tile.round();
-
-                if (**PLABELEN && shouldShow(id)) {
-                    std::string label;
-                    const std::string mode{*PLABELMODE};
-                    if (mode == "token") {
-                        if (tokenCounter < (int)labelTokens.size() && !labelTokens[tokenCounter].empty())
-                            label = labelTokens[tokenCounter];
-                        else
-                            label = fallbackTokenForVisibleIndex(tokenCounter);
-                    } else if (mode == "index") {
-                        label = std::to_string(tokenCounter + 1);
-                    } else {
-                        label = std::to_string(images[id].workspaceID);
-                    }
-
-                    const int st = resolveState(id);
-                    if (!label.empty()) {
-                        if (st == 1)
-                            renderLabel(images[id].labelTexHover, images[id].labelSizeHover, label, CHyprColor{(uint64_t)**PLCOLHOV}, **PLSCALEH, tile, std::string{*PLABELPOS}, **PLABELOX, **PLABELOY);
-                        else if (st == 2)
-                            renderLabel(images[id].labelTexFocus, images[id].labelSizeFocus, label, CHyprColor{(uint64_t)**PLCOLFOC}, **PLSCALEF, tile, std::string{*PLABELPOS}, **PLABELOX, **PLABELOY);
-                        else if (st == 3)
-                            renderLabel(images[id].labelTexCurrent, images[id].labelSizeCurrent, label, CHyprColor{(uint64_t)**PLCOLCUR}, 1.0f, tile, std::string{*PLABELPOS}, **PLABELOX, **PLABELOY);
-                        else
-                            renderLabel(images[id].labelTexDefault, images[id].labelSizeDefault, label, CHyprColor{(uint64_t)**PLCOLDEF}, 1.0f, tile, std::string{*PLABELPOS}, **PLABELOX, **PLABELOY);
-                    }
-                }
-
-                if (**PSELECTEN && tokenCounter < (int)selectionTokens.size() && !selectionTokens[tokenCounter].empty())
-                    renderLabel(images[id].selectionLabelTex, images[id].selectionLabelSize, selectionTokens[tokenCounter], CHyprColor{(uint64_t)**PSELECTCOL}, 1.0f, tile,
-                                std::string{*PSELECTPOS}, **PSELECTOX, **PSELECTOY);
-
-                tokenCounter++;
-            }
-        }
-    }
-
-    // draw borders for hover, current and focus (priority order: focus > current > hover)
-
-    // pass rounding based on state
-    const int RND_CUR = CURRENT_ROUND_SCALED;
-    const int RND_FOC = FOCUS_ROUND_SCALED;
-    const int RND_HOV = HOVER_ROUND_SCALED;
-
-    // Helper to parse border config (supports rgb/hex/gradient, with deprecated fallback)
-    auto drawBorderForID = [&](int id, const std::string& borderSpec, const std::string& deprecatedGradSpec, int roundScaled) {
-        if (id < 0)
-            return;
-        const int ix = id % SIDE_LENGTH;
-        const int iy = id / SIDE_LENGTH;
-        CBox       box{OUTER + ix * tileRenderSize.x + ix * GAPSIZE, OUTER + iy * tileRenderSize.y + iy * GAPSIZE, tileRenderSize.x, tileRenderSize.y};
-        box.scale(pMonitor->m_scale).translate(pos->value());
-        box.round();
-        const int BWIDTH = std::max(1, (int)**PBWIDTH);
-
-        // Determine which spec to use (prefer new format, fallback to deprecated)
-        std::string effectiveSpec = borderSpec.empty() ? deprecatedGradSpec : borderSpec;
-
-        // Auto-detect format: gradient vs solid color
-        if (isGradientBorderSpec(effectiveSpec)) {
-            // Render as gradient border (hyprland style)
-            const auto spec = parseGradientSpec(effectiveSpec);
-            if (spec.valid) {
-                Config::CGradientValueData grad;
-                grad.m_colors.clear();
-                grad.m_colors.push_back(spec.c1);
-                grad.m_colors.push_back(spec.c2);
-                grad.m_angle = spec.angleDeg * (float)M_PI / 180.f;
-                grad.updateColorsOk();
-                Render::GL::g_pHyprOpenGL->renderBorder(box, grad, {.round = roundScaled, .roundingPower = ROUND_PWR, .borderSize = BWIDTH});
-            }
-        } else if (!effectiveSpec.empty()) {
-            // Parse as solid color (rgb/hex format)
-            CHyprColor color;
-
-            // Handle rgb() format
-            if (effectiveSpec.find("rgb(") == 0) {
-                // Extract hex from rgb(rrggbb)
-                size_t start = effectiveSpec.find('(');
-                size_t end = effectiveSpec.find(')');
-                if (start != std::string::npos && end != std::string::npos) {
-                    std::string hexStr = effectiveSpec.substr(start + 1, end - start - 1);
-                    // Parse as hex, prepend FF for alpha
-                    color = CHyprColor{(uint64_t)std::stoull("FF" + hexStr, nullptr, 16)};
-                }
-            } else {
-                // Handle 0xAARRGGBB format directly
-                color = CHyprColor{(uint64_t)std::stoull(effectiveSpec, nullptr, 16)};
-            }
-
-            // Render as simple border
-            Render::GL::g_pHyprOpenGL->renderBorder(box, color, {.round = roundScaled, .roundingPower = ROUND_PWR, .borderSize = BWIDTH});
-        }
-    };
-
-    // Draw borders in order: hover (lowest), then current, then focus (highest priority)
-    if (hoveredID != -1 && hoveredID != openedID && hoveredID != kbFocusID)
-        drawBorderForID(hoveredID, std::string{*PBCOLHOV}, std::string{*PBGREHOV}, RND_HOV);
-    drawBorderForID(openedID, std::string{*PBCOLCUR}, std::string{*PBGRCUR}, RND_CUR);
-    if (kbFocusID != -1)
-        drawBorderForID(kbFocusID, std::string{*PBCOLFOC}, std::string{*PBGREFOC}, RND_FOC);
-}
-
-static float lerpFloat(const float& from, const float& to, const float perc) {
-    return (to - from) * perc + from;
-}
-
-static Vector2D lerp(const Vector2D& from, const Vector2D& to, const float perc) {
-    return Vector2D{lerpFloat(from.x, to.x, perc), lerpFloat(from.y, to.y, perc)};
-}
-
-void COverview::setClosing(bool closing_) {
-    closing = closing_;
-}
-
-void COverview::resetSwipe() {
-    swipeWasCommenced = false;
-}
-
-void COverview::onSwipeUpdate(double delta) {
-    m_isSwiping = true;
-
-    if (swipeWasCommenced)
-        return;
-
-    static auto* const* PDISTANCE = (Hyprlang::INT* const*)HyprlandAPI::getConfigValue(PHANDLE, "plugin:hyprexpo:gesture_distance")->getDataStaticPtr();
-
-    const float         PERC               = closing ? std::clamp(delta / (double)**PDISTANCE, 0.0, 1.0) : 1.0 - std::clamp(delta / (double)**PDISTANCE, 0.0, 1.0);
-    const auto          WORKSPACE_FOCUS_ID = closing && closeOnID != -1 ? closeOnID : openedID;
-
-    Vector2D            tileSize = (pMonitor->m_size / SIDE_LENGTH);
-
-    const auto          SIZEMAX = pMonitor->m_size * pMonitor->m_size / tileSize;
-    const auto          POSMAX  = (-((pMonitor->m_size / (double)SIDE_LENGTH) * Vector2D{WORKSPACE_FOCUS_ID % SIDE_LENGTH, WORKSPACE_FOCUS_ID / SIDE_LENGTH}) * pMonitor->m_scale) *
-        (pMonitor->m_size / tileSize);
-
-    const auto SIZEMIN = pMonitor->m_size;
-    const auto POSMIN  = Vector2D{0, 0};
-
-    size->setValueAndWarp(lerp(SIZEMIN, SIZEMAX, PERC));
-    pos->setValueAndWarp(lerp(POSMIN, POSMAX, PERC));
-}
-
-void COverview::onSwipeEnd() {
-    const auto SIZEMIN = pMonitor->m_size;
-    const auto SIZEMAX = pMonitor->m_size * pMonitor->m_size / (pMonitor->m_size / SIDE_LENGTH);
-    const auto PERC    = (size->value() - SIZEMIN).x / (SIZEMAX - SIZEMIN).x;
-    if (PERC > 0.5) {
-        close();
-        return;
-    }
-    *size = pMonitor->m_size;
-    *pos  = {0, 0};
-
-    size->setCallbackOnEnd([this](WP<Hyprutils::Animation::CBaseAnimatedVariable> thisptr) { redrawAll(true); });
-
-    swipeWasCommenced = true;
-    m_isSwiping       = false;
-}
-
-void COverview::enterSubmapIfEnabled() {
-    static auto* const* PKEYNAV = (Hyprlang::INT* const*)HyprlandAPI::getConfigValue(PHANDLE, "plugin:hyprexpo:keynav_enable")->getDataStaticPtr();
-    if (**PKEYNAV && !submapActive) {
-        // switch to a dedicated submap for hyprexpo navigation
-        g_pKeybindManager->m_dispatchers["submap"]("hyprexpo");
-        submapActive = true;
-    }
-}
-
-void COverview::resetSubmapIfNeeded() {
-    if (submapActive) {
-        g_pKeybindManager->m_dispatchers["submap"]("reset");
-        submapActive = false;
-    }
-}
