+++ scripts/run-nested.sh
@@ -1,0 +1,126 @@
+#!/usr/bin/env bash
+set -euo pipefail
+
+# Launch a nested Hyprland session that loads the local hyprexpo.so,
+# so you can test changes without restarting your main session.
+
+SCRIPT_DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
+REPO_ROOT="$(cd "$SCRIPT_DIR/.." && pwd)"
+BUILD_DIR="${XDG_CACHE_HOME:-$HOME/.cache}/hyprexpo-plus"
+SO="${HYPREXPO_DEV_SO:-$BUILD_DIR/hyprexpo.so}"
+CONF="${XDG_CACHE_HOME:-$HOME/.cache}/hyprexpo-dev.conf"
+
+mkdir -p "$(dirname "$CONF")" "$(dirname "$SO")"
+echo "[run-nested] Building local plugin at $SO"
+make -C "$REPO_ROOT" all TARGET="$SO"
+
+cat > "$CONF" <<EOF
+monitor=,preferred,auto,auto
+
+debug {
+  disable_logs = false
+}
+
+cursor {
+  no_hardware_cursors = true
+}
+
+# load local build
+plugin = $SO
+
+plugin {
+  hyprexpo {
+    # layout + visuals
+    columns = 3
+    gaps_in = 20
+    bg_col = rgb(101010)
+    workspace_method = center current
+    skip_empty = 0
+
+    # borders (hypr-style gradient, thicker to showcase)
+    border_style = hyprland
+    border_width = 4
+    border_color_current = rgb(66ccff)
+    border_color_focus   = rgb(ffcc66)
+
+    # keyboard nav
+    keynav_enable = 1
+    keynav_wrap_h = 1
+    keynav_wrap_v = 1
+    keynav_reading_order = 0
+
+    # labels (numbers): smaller font, rounded background bubble
+    label_enable = 1
+    label_font_size = 12
+    label_position = bottom-right
+    label_offset_x = 8
+    label_offset_y = 8
+    label_show = hover+focus
+    label_color_default = rgb(ffffff)
+    label_color_hover   = rgb(72ff7a)
+    label_color_focus   = rgb(ffcc66)
+    label_color_current = rgb(66ccff)
+    label_scale_hover = 1.0
+    label_scale_focus = 1.2
+    label_bg_enable = 1
+    label_bg_color = rgba(000000cc)
+    label_bg_rounding = 999  # fully rounded bubble
+    label_padding = 6
+
+    # outer margin around the grid to demo spacing from screen edge
+    gaps_out = 20
+
+    # demo hyprland style gradient borders
+    border_grad_current = rgba(33ccffee) rgba(00ff99ee) 45deg
+    border_grad_focus   = rgba(ffdd44ee) rgba(22aaffee) 30deg
+  }
+}
+
+# toggle with an unmodified function key to avoid host grabs
+bind = , F10, exec, hyprctl dispatch hyprexpo:expo toggle
+
+# nested-session test controls
+bind = SUPER, Return, exec, kitty
+bind = SUPER, Q, killactive
+bind = SUPER SHIFT, Q, exit
+bind = SUPER, 1, workspace, 1
+bind = SUPER, 2, workspace, 2
+bind = SUPER, 3, workspace, 3
+bind = SUPER, 4, workspace, 4
+bind = SUPER, 5, workspace, 5
+bind = SUPER, 6, workspace, 6
+bind = SUPER, 7, workspace, 7
+bind = SUPER, 8, workspace, 8
+bind = SUPER, 9, workspace, 9
+bind = SUPER SHIFT, 1, movetoworkspace, 1
+bind = SUPER SHIFT, 2, movetoworkspace, 2
+bind = SUPER SHIFT, 3, movetoworkspace, 3
+bind = SUPER SHIFT, 4, movetoworkspace, 4
+bind = SUPER SHIFT, 5, movetoworkspace, 5
+bind = SUPER SHIFT, 6, movetoworkspace, 6
+bind = SUPER SHIFT, 7, movetoworkspace, 7
+bind = SUPER SHIFT, 8, movetoworkspace, 8
+bind = SUPER SHIFT, 9, movetoworkspace, 9
+
+# submap for keyboard nav (the plugin auto-enters this when open)
+submap = hyprexpo
+  bind = , left,  exec, hyprctl dispatch hyprexpo:kb_focus left
+  bind = , right, exec, hyprctl dispatch hyprexpo:kb_focus right
+  bind = , up,    exec, hyprctl dispatch hyprexpo:kb_focus up
+  bind = , down,  exec, hyprctl dispatch hyprexpo:kb_focus down
+  bind = , return, exec, hyprctl dispatch hyprexpo:kb_confirm
+  bind = , 1, exec, hyprctl dispatch hyprexpo:kb_selectn 1
+  bind = , 2, exec, hyprctl dispatch hyprexpo:kb_selectn 2
+  bind = , 3, exec, hyprctl dispatch hyprexpo:kb_selectn 3
+  bind = , 4, exec, hyprctl dispatch hyprexpo:kb_selectn 4
+  bind = , 5, exec, hyprctl dispatch hyprexpo:kb_selectn 5
+  bind = , 6, exec, hyprctl dispatch hyprexpo:kb_selectn 6
+  bind = , 7, exec, hyprctl dispatch hyprexpo:kb_selectn 7
+  bind = , 8, exec, hyprctl dispatch hyprexpo:kb_selectn 8
+  bind = , 9, exec, hyprctl dispatch hyprexpo:kb_selectn 9
+  bind = , 0, exec, hyprctl dispatch hyprexpo:kb_selectn 0
+submap = reset
+EOF
+
+echo "[run-nested] Launching nested Hyprland with $CONF"
+exec env WLR_BACKENDS=wayland WLR_RENDERER=pixman WLR_NO_HARDWARE_CURSORS=1 HYPRLAND_NO_LOGO=1 Hyprland -c "$CONF"
